fixed components names
