+++ src/components/GeoAstroEvents/GeoAstroEvents.jsx
@@ -1,9 +1,10 @@
 
 import Earthquake from "./Earthquake"
-import LunarEclipse from "./LunarEclipse";
-import SolarEclipse from "./LunarEclipse";
+
 import Meteorshowers from "./Meteorshowers";
+import SolarEclipse from "./SolarEclipse";
 import Volcanoes from "./Volcanoes";
+import LunarEclipse from "./LunarEclipse"
 function GeoAstroEvents() {
 
   return (
@@ -15,13 +16,12 @@ function GeoAstroEvents() {
           <Earthquake />
           <h1>Lunar Eclipses</h1>
           <h3>25 YRS. Past-Future</h3>
-
           <LunarEclipse />
+
           <h1>Solar Eclipses</h1>
           <h3>25 YRS. Past-Future</h3>
 
           <SolarEclipse />
-
           <h1>Major Volcanic Activity</h1>
           <h3>25 YRS. Past-Potential-Future</h3>
           <Volcanoes />
